DPL: check for state transition updates also in input-proxy busy loop
